feat: VoteSearchProperty 분리
