+++ Projects/Feature/Vote/Sources/VoteSearch/VoteSearch.swift
@@ -22,8 +22,7 @@ struct VoteSearch {
     var searchReducer: SSSearchReducer<VoteSearchProperty>.State
 
     init() {
-      let helper = VoteSearchProperty()
-      _helper = Shared(helper)
+      _helper = Shared(.init())
       searchReducer = .init(helper: _helper)
     }
   }
@@ -86,59 +85,5 @@ struct VoteSearch {
         return .none
       }
     }
-  }
-}
-
-// MARK: - VoteSearchItem
-
-struct VoteSearchItem: SSSearchItemable {
-  var id: Int
-  var title: String
-  var firstContentDescription: String?
-  var secondContentDescription: String?
-}
-
-// MARK: - VoteSearchProperty
-
-struct VoteSearchProperty: SSSearchPropertiable {
-  typealias item = VoteSearchItem
-
-  var textFieldPromptText: String
-  var prevSearchedNoContentTitleText: String
-  var prevSearchedNoContentDescriptionText: String
-  var noSearchResultTitle: String
-  var noSearchResultDescription: String
-  var iconType: SSSearchIconType
-
-  mutating func searchItem(by _: String) {}
-  mutating func deletePrevItem(prevItemID _: Int) {}
-
-  var textFieldText: String = ""
-  var prevSearchedItem: [VoteSearchItem] = []
-  var nowSearchedItem: [VoteSearchItem] = []
-
-  init(
-    textFieldPromptText: String,
-    prevSearchedNoContentTitleText: String,
-    prevSearchedNoContentDescriptionText: String,
-    noSearchResultTitle: String,
-    noSearchResultDescription: String,
-    iconType: SSSearchIconType
-  ) {
-    self.textFieldPromptText = textFieldPromptText
-    self.prevSearchedNoContentTitleText = prevSearchedNoContentTitleText
-    self.prevSearchedNoContentDescriptionText = prevSearchedNoContentDescriptionText
-    self.noSearchResultTitle = noSearchResultTitle
-    self.noSearchResultDescription = noSearchResultDescription
-    self.iconType = iconType
-  }
-
-  init() {
-    textFieldPromptText = "찾고 싶은 투표를 검색해보세요"
-    prevSearchedNoContentTitleText = "어떤 투표를 찾아드릴까요?"
-    prevSearchedNoContentDescriptionText = "궁금하신 것들의 키워드를\n검색해볼 수 있어요"
-    noSearchResultTitle = "원하는 검색 결과가 없나요?"
-    noSearchResultDescription = "SomeTing"
-    iconType = .vote
   }
 }
